requirements.txt
flask
gunicorn
flask-cors
opencv-python-headless
numpy
mediapipe==0.10.14
google-generativeai==0.5.4